we can regard the api changed as new api, so version is v1
